requirements.txt
clip-by-openai
nltk
numpy
Pillow
torch      
torchvision
tqdm
transformers
opencv-python
